darwinbuild now exports the RC_BuildRoot variable
You use it by prepending '${RC_BuildRoot}' to paths to files
installed by a darwinbuild root that were previously absolute.

This environment variable, usable in both make-based build systems
and Xcode project-based systems, is equivalent to the path to what
would be the root of the chroot, if the chroot is not being used.
If the chroot *is* being used, it will not be defined, therefore
evaluate to an empty string, and previously-absolute paths would
still be absolute.