Modify darwinbuild so that it can build as a non-root user
This makes it possible for code signing to occur during the build,
as if the user built the project using the Xcode user interface.
If using the chroot, root privileges will still be required.

Also removed NFS support, as not only does that require the use
of the root account, it is also thoroughly deprecated.