Fix path manipulation
The previous algorithm would not work on a stock macOS
system (it depended on coreutils, which I happened to have
installed at the time), and it would also cause failures if
the build root image was not mounted (which therefore broke
mounting the image to begin with). Both problems should
now be fixed.