Remove use of deprecated redo_prebinding tool
Prebinding became well and truly obsolete with the change to
Intel processors during the tenure of Mac OS X Tiger. There
is no reason why we should still have logic to support it.

Also, redo_prebinding is not included in Xcode 11; it is only
present in the command line tools download. This causes a dialog
to appear if the command line tools are not installed.