Vendor Tcl
Future versions of macOS will remove Tcl from the in-box
SDK. So that darwinxref does not break, I must vendor a
copy of libtcl8.6.dylib. I have chosen to copy this library
from the tcl-tk Homebrew formula, which must now be present
on the computer you are building darwinbuild on.