Add ~n suffixes before the filename extension
This ensures that build log files will always in '.log',
which enables the Finder and other applications to automaticllly
recognize them as text files. Previously, the build number was
appended at the end of the extension, resulting in filenames like
"AvailabilityVersions-45.5.log~5". The new filename would be
"AvailabilityVersions-45.5~5.log".

I also made the same change to the .trace files emitted
when you pass the -logdeps flag to darwinbuild, which now
always end in '.deps.log'.