Rewrite darwinbuild-codesign tool
This tool is no longer intended to be invoked manually.
It is now designed to be run by darwinbuild as a post-
processing step after a build. It now locates its settings
plist from within the DSTROOT, and no longer requires the
path to the corresponding SRCROOT. (This was only needed
to locate paths to *.entitlements files. I now know how to
tell codesign to preserve the existing entitlements while
re-signing everything else.)